RavenDB-12295 Fix failing tests
Use the nonPersistentFlags instead of the Persistent ones to detect if 'FromReplication'.
