+++ src/Raven.Server/Documents/ConflictsStorage.cs
@@ -784,7 +784,8 @@ namespace Raven.Server.Documents
                     context.GetLazyString(mergedChangeVector),
                     latestConflict.LastModified.Ticks,
                     changeVector,
-                    latestConflict.Flags).Etag;
+                    latestConflict.Flags, 
+                    NonPersistentDocumentFlags.None).Etag;
             }
 
             return collectionName;
